# lib/api-zod/src/generated/types/chatCompletionChoice.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface ChatCompletionChoice {
  index: number;
  message: ChatMessage;
  /** @nullable */
  finish_reason: string | null;
}