# lib/api-zod/src/generated/types/chatCompletionResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatCompletionChoice } from './chatCompletionChoice';
import type { ChatCompletionUsage } from './chatCompletionUsage';

export interface ChatCompletionResponse {
  id: string;
  object: string;
  created: number;
  model: string;
  choices: ChatCompletionChoice[];
  usage: ChatCompletionUsage;
}