# lib/api-zod/src/generated/types/memoryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MemoryInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  category: string;
}