# lib/api-zod/src/generated/types/modelActionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ModelActionInput {
  /** @minLength 1 */
  modelId: string;
}